--- codeception.yml
@@ -1,17 +1,0 @@
-suites:
-  unit:
-    path: .
-    actor: UnitTester
-    modules:
-      enabled:
-      # add more modules here
-      - Asserts
-settings:
-  shuffle: true
-  lint: true
-paths:
-  tests: tests
-  output: tests/_output
-  support: tests/_support
-  data: tests
-bootstrap: _bootstrap.php
