--- tests/unit.suite.yml
@@ -1,4 +1,0 @@
-actor: UnitTester
-modules:
-  enabled:
-  - \Helper\Unit
